+++ app/src/main/java/fi/kroon/vadret/presentation/weatherforecast/model/WeatherForecastHeadlineModel.kt
@@ -1,5 +1,7 @@
 package fi.kroon.vadret.presentation.weatherforecast.model
 
 data class WeatherForecastHeadlineModel(
-    val headline: Int?
+    val headline: Int?,
+    val windSpeed: Double?,
+    val windDirection: Double?
 ) : BaseWeatherForecastModel